kci_test: Refactor LAVA XMLRPC API handling
- Extract common parts from the LAVA API implementation.
- Create lab.lava package
- Move XMLRPC based communication to lab.lava.lava_xmlrpc.
The redesign is a preparation for another implementing other means of
communication with LAVA labs.
